refactor: divide game datum into game-data.json
